requirements.txt
aiohttp==3.8.3
aiohttp-socks==0.7.1
aiorpcX==0.22.1
aiosignal==1.3.1
anyio==3.6.2
anytree==2.8.0
apparmor==3.1.2
appdirs==1.4.4
apsw==3.40.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asn1crypto==1.5.1
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autocommand==2.2.2
autopage==0.5.1
Babel==2.11.0
backcall==0.2.0
Beaker==1.12.0
beautifulsoup4==4.11.1
binwalk==2.3.3
bitstring==3.1.9
black==22.12.0
bleach==5.0.1
blinker==1.5
borgbackup==1.2.3
borgmatic==1.7.6
Brlapi==0.8.4
Brotli==1.0.9
brotlicffi==1.0.9.2
btrfsutil==6.1.2
CacheControl==0.12.11
cchardet==2.1.7
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
cliff==4.1.0
clint==0.5.1
cmd2==2.4.2
colorama==0.4.6
comm==0.1.2
commonmark==0.9.1
configobj==5.1.0.dev0
contextlib2==21.6.0
cryptography==39.0.0
css-parser==1.0.8
cssselect==1.2.0
Cython==0.29.33
debtcollector==2.5.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.6
distro==1.8.0
dnspython==2.2.1
docopt==0.6.2
docutils==0.19
dogpile.cache==1.1.8
ecdsa==0.18.0
Electrum==4.3.2
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
feedparser==6.0.10
fido2==1.1.0
filelock==3.8.2
Flask==2.2.2
frozenlist==1.3.3
future==0.18.2
guake==3.9.1.dev0
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2020.1.16
html5-parser==0.4.10
html5lib==1.1
hyperframe==6.0.1
idna==3.4
ifaddr==0.2.0
imbox==0.9.8
importlib-metadata==5.0.0
inflate64==0.3.1
inflect==6.0.2
ipykernel==6.20.2
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
iso8601==1.1.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jaraco.context==4.2.0
jaraco.functools==3.5.2
jaraco.text==3.11.0
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
json5==0.9.12
jsonpatch==1.32
jsonpointer==2.3
jsonrpclib-pelix==0.4.3.2
jsonschema==4.17.3
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==8.0.1
jupyter_core==5.1.5
jupyter_packaging==0.12.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab==3.5.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.19.0
kaitaistruct==0.10
keyring==23.11.0
keystoneauth1==5.1.1
ldap3==2.9
lensfun==0.3.3
LibAppArmor==3.1.2
libfdt==1.6.1
libvirt-python==9.0.0
louis==3.24.0
lxml==4.9.2
Mako==1.2.4
mallard-ducktype==1.0.2
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mechanize==0.4.8
mistune==2.0.4
mitmproxy==9.0.1
mitmproxy_wireguard==0.1.20
more-itertools==9.0.0
msgpack==1.0.4
multidict==6.0.3
multivolumefile==0.2.3
munch==2.5.0
mutagen==1.45.1
mypy-extensions==0.4.3
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
netaddr==0.8.0
netifaces==0.11.0
netsnmp-python==1.0a1
nftables==0.1
notebook==6.5.2
notebook_shim==0.2.2
nspektr==0.4.0
numpy==1.24.0
openstacksdk==0.103.0
ordered-set==4.1.0
os-client-config==2.1.0
os-service-types==1.7.0
osc-lib==2.6.2
oslo.config==9.1.0
oslo.i18n==5.1.0
oslo.serialization==5.0.0
oslo.utils==6.1.0
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pathspec==0.11.0
pbkdf2==1.3
pbr==5.11.1
pdftotext==2.2.2
peewee==3.13.3
pep517==0.13.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip-upgrade==0.0.6
platformdirs==2.6.2
ply==3.11
prettytable==3.6.0
progress==1.6
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.4
ptyprocess==0.7.0
publicsuffix2==2.20191221
pure-eval==0.2.2
pwquality==1.4.5
py7zr==0.20.2
pyaes==1.6.1
pyasn1==0.4.8
pybcj==1.0.1
pycairo==1.23.0
pychm==0.8.6
pycparser==2.21
pycryptodome==3.16.0
pycryptodomex==3.12.0
pydantic==1.10.4
Pygments==2.14.0
PyGObject==3.42.2
pykwalify==1.8.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyperclip==1.8.2
pyppmd==1.0.0
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQt6==6.4.0
PyQt6-sip==13.4.0
PyQt6-WebEngine==6.4.0
PyQtWebEngine==5.15.6
pyrsistent==0.19.3
pyscard==2.0.5
PySocks==1.7.1
python-cinderclient==9.2.0
python-dateutil==2.8.2
python-glanceclient==4.1.0
python-json-logger==2.0.4
python-keystoneclient==5.0.1
python-novaclient==18.2.0
python-openstackclient==6.0.0
python-socks==2.0.3
pytoml==0.1.21
pytz==2022.7
pyusb==1.2.1
pyxdg==0.28
PyYAML==6.0
pyzmq==24.0.1
pyzstd==0.15.3
qrcode==7.3.1
rados==2.0.0
rbd==2.0.0
regex==2022.10.31
requests==2.28.1
requests-file==1.5.1
requests-unixsocket==0.3.0
requestsexceptions==1.4.0
resolvelib==0.9.0
retrying==1.3.3
rfc3986==2.0.0
rfc3987==1.3.8
rgw==2.0.0
rich==13.2.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
SecretStorage==3.3.3
Send2Trash==1.8.0
setproctitle==1.3.2
sgmllib3k==1.0.0
shiboken2==5.15.8
simplejson==3.18.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spotipy==2.16.1
stack-data==0.6.2
stevedore==4.1.1
TBB==0.2
team==1.0
tenacity==8.1.0
terminado==0.17.1
terminator==2.1.2
testpath==0.6.0
texttable==1.6.7
tinycss2==1.2.1
tldextract==3.4.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
traitlets==5.8.1
trove-classifiers==2023.1.20
typing_extensions==4.4.0
uc-micro-py==1.0.1
unrardll==0.1.5
urllib3==1.26.12
urwid==2.1.2
validate==5.1.0.dev0
validate-pyproject==0.11
warlock==2.0.1
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wrapt==1.14.1
wsproto==1.2.0
yarl==1.8.2
youtube-dl==2021.12.17
yubikey-manager==4.0.9
zeroconf==0.39.4
zipfile-deflate64==0.2.0
zipp==3.11.0
zstandard==0.19.0
